% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire data set.}
\format{A data frame with 15 rows and 2 variables:
\describe{
  \item{DISTANCE}{distance in cm}
  \item{DAMAGE}{damage ammount}
  ...
}}
\source{
\url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
fire
}
\description{
fire data
}
\keyword{datasets}
