% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbinpois.R
\name{logbinpois}
\alias{logbinpois}
\title{logbinpois}
\usage{
logbinpois(theta1, theta2)
}
\arguments{
\item{theta2}{}
}
\value{
a list of vectors
}
\description{
logbinpois
}
\examples{
theta1=10, theta 2= 20

This function is a function that you can use
in the maxlikg2 function. It takes thetas and evaluates theta one with dbinom and theta 2 with dpois and creates a vector of values.


}
