% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myboot2.R
\name{myboot2}
\alias{myboot2}
\title{My Boot2 Function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
My Boot2 Function
}
\examples{
iter= 100000, x= sam


This function creates a histogram of bootstrap statistics for a sample with a number of iterations within a specific confidence interval

}
