% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyNCurve.R
\name{myncurve}
\alias{myncurve}
\title{My N Curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{sigma}{}
}
\value{
a list of vectors and a graph of probaiblities
}
\description{
My N Curve
}
\examples{

This function creates a graph of probabilies based on the provided mu, sigma, and gives the option to shade in a region defined by a

}
