% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{Mypvalue}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{alpha}{}
}
\value{
a graph and value for denisity of t test and p-values with a value for p-values
}
\description{
Mypvalue
}
\examples{
mypvalue(t0=1.23, xmax= 4, n=33, alpha=0.05)

This function creates a p-value based on a provided t-value, n, and a confidence interval
and also creates a barplot to show the rejection region for t-values.


}
