% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyBin.R
\name{mybin}
\alias{mybin}
\title{My Bin Function}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{p}{}
}
\value{
a list of numbers
}
\description{
My Bin Function
}
\examples{
x:10, mybin(x)
This function estimates the probability with 100 iterations, n=10, and a p value of 0.5

}
