% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My First Lab Function.R
\name{myplot}
\alias{myplot}
\title{Myplot Functions}
\usage{
myplot(x)
}
\arguments{
\item{x}{a vector of double values}
}
\value{
a vector of doubles
}
\description{
This Function Calculates the quadratic curve for the spruce df set of data
}
\details{
Has the coefficients to create the plot and equation for the quadratic equation in the spruce data
}
\examples{
x:2; myplot(x)
}
